Updated script to measure energy.  Also refactored to ease adding more run parameters (e.g. prefetchers).  Also enabled measuring user metrics.
Former-commit-id: 6e1e8e5fc8e13206b466e63f78457bad8f8bad4a [formerly 7e16d4187e61129493e7b89839b475e370505a77] [formerly 17489296ed3c5bdadb4d9b7cef2435498bf1b903 [formerly 1b8ef14b73564832050e1f4dbcfdd47687b73a79]]
Former-commit-id: d7a56bd9c539d8329b292502dea26ef291fb20ee [formerly f29535eeac1fe914904e12abca3dc00449e3f800]
Former-commit-id: e4893c3fd4106dbd7c77a517e13c355323a26f2b

+++ format2cape.sh
@@ -6,7 +6,7 @@ source const.sh
 
 nb_args=$#
 
-if [[ $nb_args -lt 4 ]]
+if [[ $nb_args -lt 1 ]]
 then
 	echo "ERROR! Invalid arguments (need: CLS res's folder, real machine name, variants)."
 #	echo "ERROR! Invalid arguments (need: CLS res's folder)."
@@ -19,9 +19,10 @@ source pcr_metrics.sh
 
 # Get codelet's res folder
 cls_res_folder=$( readlink -f "$1" )
-real_machine_name="$2"
-variants="$3"
-num_cores="$4"
+#real_machine_name="$2"
+real_machine_name="$(hostname)"
+#variants="$3"
+#num_cores="$4"
 
 codelet_folder=$( echo $cls_res_folder | sed 's:/cls_res_.*::' )
 #machine_name=$( echo $cls_res_folder | sed -n -E 's/(.*cls_res_)(.*)(_[0-9]+)/\2/p' )
@@ -68,14 +69,125 @@ REPETITIONS_COL=7
 NUMCORE_COL=5
 
 cpu_generation="$(cat $cls_res_folder/uarch)"
+energy_unit="$(cat $cls_res_folder/energy_units)"
 
 
 gen_codelet_mach_info () {
     local nr="$1"
     local of="$2"
-    echo "application.name"${DELIM}"batch.name"${DELIM}"code.name"${DELIM}"codelet.name"${DELIM}"binary_loop.id"${DELIM}"decan_variant.name"${DELIM}"machine.name"${DELIM}"real.machine.name"> $of
-    yes $(echo "$application_name"${DELIM}"$batch_name"${DELIM}"$code_name"${DELIM}"$codelet_name"${DELIM}"$binary_loop_id"${DELIM}"$variant"${DELIM}"$machine_name"${DELIM}"$real_machine_name") | head -n $nr >> $of
+    echo "application.name"${DELIM}"batch.name"${DELIM}"code.name"${DELIM}"codelet.name"${DELIM}"binary_loop.id"${DELIM}"machine.name"${DELIM}"real.machine.name"> $of
+    yes $(echo "$application_name"${DELIM}"$batch_name"${DELIM}"$code_name"${DELIM}"$codelet_name"${DELIM}"$binary_loop_id"${DELIM}"$machine_name"${DELIM}"$real_machine_name") | head -n $nr >> $of
 }
+
+gen_filler_info() {
+   local nr="$1"
+   local of="$2"
+   echo "decan_experimental_configuration.instance_id"${DELIM}"decan_experimental_configuration.nb_threads"${DELIM}"Timestamp"${DELIM}"cpu.generation"${DELIM}"energy.unit"${DELIM}"Expr Timestamp"${DELIM}"TS#"${DELIM}"Expr TS#" > $of
+   yes $(echo "$instance_id"${DELIM}"$nb_threads"${DELIM}"$ClsTimestamp"${DELIM}"$cpu_generation"${DELIM}"$energy_unit"${DELIM}"$ExprTimestamp"${DELIM}"$cls_timestamp_val"${DELIM}"$run_timestamp_val") | head -n $nr >> $of
+}
+
+counter_value_files=$(find $cls_res_folder -name 'counter_values.csv')
+counter_name_files=$(find $cls_res_folder -name 'counter_names.csv')
+counter_names=$( cat $counter_name_files | uniq)
+counter_values=$( cat $counter_value_files )
+
+if [[ $(echo "$counter_names" | wc -l) != "1" ]]
+then
+    echo "Unexpected corrupted counter collection, exiting"
+    exit -1
+fi
+
+tmprep=$(mktemp -d --tmpdir=$cur_dir tmp.XXXXXXXXXX)
+
+echo "$counter_names" > $tmprep/counters.csv
+echo "$counter_values" >> $tmprep/counters.csv
+
+cpi_iteration_rep_value_files=$(echo "$counter_value_files" |sed 's/counter_values/cpi_values/g')
+cpi_iteration_rep_name_files=$(echo "$counter_value_files" |sed 's/counter_values/cpi_names/g')
+cpi_iteration_rep_names=$( cat $cpi_iteration_rep_name_files | uniq )
+cpi_iteration_rep_values=$( cat $cpi_iteration_rep_value_files )
+echo "$cpi_iteration_rep_names" > $tmprep/cpi_iteration_rep.csv
+echo "$cpi_iteration_rep_values" >> $tmprep/cpi_iteration_rep.csv
+
+pgm_metrics_files=$(echo "$counter_value_files" |sed 's/counter_values/pgm_metrics/g')
+pgm_names=$(head -1 -q $pgm_metrics_files|uniq)
+pgm_values=$(tail -n +2 -q $pgm_metrics_files)
+echo "$pgm_names" > $tmprep/pgm.csv
+echo "$pgm_values" >> $tmprep/pgm.csv
+
+
+# <same repeated machine/filler info> below
+num_rows=$(echo "$counter_values" | wc -l)
+gen_codelet_mach_info $num_rows $tmprep/codelet_mach_info.csv
+gen_filler_info $num_rows $tmprep/filler_info.csv # including timestamp and cpu generation info
+
+# <per run setting info> below
+runinfo=$(echo "$counter_value_files" |sed 's|'$cls_res_folder'||g')  # Used '|' as delimiter as cls_res_folder contains slash
+runinfo=$(echo "$runinfo" |sed 's|/counter_values\.csv||g') 
+runinfo_values=$(echo "$runinfo" |sed 's|/[^_]*_|,|g'|sed 's|^,||g')  # get rid of <setting name>_, then leading comma
+
+runinfo_names=$(echo "$runinfo" |sed 's|_[^/]*|,|g' |sed 's|/||g' | sed 's|,$||g')  # get rid of _<setting value> , then remove slashes and last comma
+# convert name using map defined in const.sh
+for m in ${!nameMap[@]};
+do
+    runinfo_names=$(echo "$runinfo_names" |sed 's|'$m'|'${nameMap[$m]}'|g')
+done
+
+#echo "$runinfo"
+#echo "$runinfo_values"
+#echo "$runinfo_names"
+
+echo "$runinfo_names" |uniq > $tmprep/runinfo.csv
+echo "$runinfo_values" >> $tmprep/runinfo.csv
+ 
+#Finally, get the stan report (if any)
+# Pickup the variant_... from runinfo and use that to construct the path to stan report
+stan_infiles=$(echo "$runinfo" | sed 's|.*/variant_\([^/]*\)/.*|'$cls_res_folder/binaries/${codelet_name}'_\1.stan_full.csv|g')
+stan_names=$(head -1 -q $stan_infiles|uniq)
+# Convert the names to make various characters becoming underscore
+stan_names=$(echo $stan_names |sed "s/[ :-]/_/g"|sed "s/_\+/_/g"|sed "s/\[/(/g"|sed "s/\]/)/g") # Also converted various naming of stan metrics to follow final format
+stan_values=$(tail -n +2 -q $stan_infiles)
+echo "$stan_names" > $tmprep/stan.csv
+echo "$stan_values" >> $tmprep/stan.csv
+
+echo Extracting stan columns only specified in stan metric file: ${STAN_METRICS_FILE}
+IFS=$'\n' need_stan_cols=($(cat ${STAN_METRICS_FILE}|sed "s/,/\n/g"))
+IFS=$'\n' stan_cols_in_file=($(head -1 $tmprep/stan.csv|sed "s/,/\n/g")) # Also converted various naming of stan metrics to follow final format
+declare -A stan_col_index
+si=0
+while [ $si -lt ${#stan_cols_in_file[@]} ]
+do
+#    echo ${stan_cols_in_file[$si]} is $si
+    (( stan_col_index[${stan_cols_in_file[$si]}]=si+1 ))  # indendedly using the si+1 because cut indices starts from 1
+    (( si++ ))
+done
+#echo "TRIMMING STAN FLIE"
+#Use the index to find out the indices of needed stan cols
+need_indices=()
+echo > $tmprep/stan_trimmed.csv
+for sc in ${need_stan_cols[@]}
+do
+    idx=${stan_col_index[$sc]}
+#    echo $sc is $idx
+    if [ x"$idx" != x ]; then
+	# Idx not empty
+	paste -d${DELIM} $tmprep/stan_trimmed.csv <(cut -d${DELIM} -f$idx $tmprep/stan.csv) > $tmprep/tmp.csv
+	mv $tmprep/tmp.csv $tmprep/stan_trimmed.csv
+    fi
+done 
+sed -i "s/^,//g" $tmprep/stan_trimmed.csv # Remove the leading comma
+
+
+# format is pretty much:
+#   <same repeated machine/filler info> <per run setting obtained from run_info> <counters> <stan data>
+paste -d${DELIM} $tmprep/codelet_mach_info.csv $tmprep/filler_info.csv $tmprep/runinfo.csv $tmprep/pgm.csv $tmprep/cpi_iteration_rep.csv $tmprep/counters.csv $tmprep/stan_trimmed.csv > $cape_file
+
+rm -R $tmprep
+
+
+exit
+# Old code below to be deleted later
+
 
 # Get variant and frequency lists
 for variant in "$cls_res_folder"/data_*/memload_*/freq_*/variant_*
@@ -92,18 +204,24 @@ cur_dir="$PWD"
 #mkdir -p $tmprep
 
 
-tmprep=$(mktemp -d --tmpdir=$cur_dir tmp.XXXXXXXXXX)
+
 
 
 for variant in $variants
 do
 	# Building the stan metrics section
 	stan_infile="$cls_res_folder/binaries/${codelet_name}_${variant}.stan_full.csv"
-#	cp $stan_infile /tmp/xxx
-	cat $stan_infile | sed -n '1p;11p'  | tr ${DELIM} '\n' | tr "=" ":"  > $tmprep/stanh.csv			
-#	cp $tmprep/stanh.csv /tmp/yyy
-	cat $stan_infile | sed -n '2p;12p'  | tr ${DELIM} '\n'               > $tmprep/stanv.csv
-#	cp $tmprep/stanv.csv /tmp/zzz
+	if [ -f $stan_infile ]; then
+	    #	cp $stan_infile /tmp/xxx
+	    cat $stan_infile | sed -n '1p;11p'  | tr ${DELIM} '\n' | tr "=" ":"  > $tmprep/stanh.csv			
+	    #	cp $tmprep/stanh.csv /tmp/yyy
+	    cat $stan_infile | sed -n '2p;12p'  | tr ${DELIM} '\n'               > $tmprep/stanv.csv
+	    #	cp $tmprep/stanv.csv /tmp/zzz
+	else
+	    # Create dummy empty file so below will have all metrics becoming zero
+	    touch $tmprep/stanh.csv $tmprep/stanv.csv
+	fi
+
 	cat $tmprep/stanh.csv | sed 's/ /_/g' | sed 's/,//g' | sed 's/\[/(/g'	| sed 's/\]/)/g' | sed 's/-/_/g' | sed 's/\.//g' | sed 's/:/_/g' | sed 's/__/_/g' | sed 's/__/_/g' | sed 's/__/_/g'| sed 's/__/_/g'	|sed 's/__/_/g' > $tmprep/tmp.csv
 #	cp $tmprep/tmp.csv /tmp/aaa
 	mv $tmprep/tmp.csv $tmprep/stanh.csv
@@ -162,6 +280,16 @@ else
     # use cpi instead as we may not enable counters
     frequency_list=$(ls $cls_res_folder/cpis/cpi_*kHz.csv | sed 's:.*MBs_::;s:kHz.csv::' | sort --uniq)
 
+    
+    for path_to_counter in $(find $cls_res_folder -name 'counter_values.csv')
+    do
+	counter_dir=$(dirname $path_to_counter)
+	run_info=${counter_dir#$cls_res_folder}
+	echo ri $run_info
+	echo $run_info |sed 's/\/[^_]*_/'${DELIM}'/g'  # remove the ..._ part and replaced by ${DELIM}
+    done
+	
+
     for variant in $variants
     do
 	#	cp $tmprep/stan_report_${variant}.csv /tmp/tt2
@@ -189,7 +317,10 @@ else
 		(IFS=${DELIM}; echo "${converted_metrics[*]}")           > $tmprep/counters.csv
 		#tail -n +2 $infile | cut -d${DELIM} -f 6-$ncols | tr ';' ',' >> $tmprep/counters.csv
 		#		cat $infile | grep $codelet_name | sed '$ d' | cut -d';' -f 6-$ncols | tr ';' ',' >> $tmprep/counters.csv
-		cat $infile | grep $codelet_name | sed '$ d' | cut -d${DELIM} -f ${BEGIN_COUNTER_DATA_COLS}-$ncols  >> $tmprep/counters.csv
+#		cat $infile | grep $codelet_name | sed '$ d' | cut -d${DELIM} -f ${BEGIN_COUNTER_DATA_COLS}-$ncols  >> $tmprep/counters.csv
+		# Slightly diff implmeentation than above line which assume last line is the extra line 
+		# with codelet name, which will fail if we ignore loop detection error and don't dump assembly
+		cat $infile | grep ${codelet_name}${DELIM} | cut -d${DELIM} -f ${BEGIN_COUNTER_DATA_COLS}-$ncols  >> $tmprep/counters.csv
 		nrows=$(tail -n +2 $tmprep/counters.csv | wc -l)
 		# Merging the stan and counters sections
